requirements.txt
# base packages
numpy
scipy
numexpr

# graphic
matplotlib

# tests
pyhamcrest
pytest
